Otimizando camdada de persistencia

+++ src/main/java/com/sceapi/model/Turma.java
@@ -18,7 +18,7 @@ public class Turma {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Double id;
+	private Long id;
 	
 	@Column(nullable = false)
 	private Date dataDeInicio;
@@ -34,11 +34,11 @@ public class Turma {
 	@JoinColumn
 	private List<Matricula> matricula;
 
-	public Double getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Double id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
